Added Admin Profile page, Added Backend for View Applicants and View Applications (#26)
Faced a lot of challenges in the AdminProfile page, as in our database schema email is a primary key, in table users. so we cannot update in easily. So finally we decided to keep the email to be immutable. a

Added backend code for View Application , to fetch the list of scholarships from the database , and backend code for View applicants.

Updated the App.js files as it had a few discrepancies.
Updated the Navbar.jsx file.
Updated ListofScholarships.jsx
Updated ViewApplicant.jsx file
Updated Userroute.js, Scholarshiproutes.js